requirements.txt
djangorestframework==3.15.2
faker==37.0.2
pip-chill==1.0.3